requirements.txt
numpy==2.2.4
opencv-python==4.11.0.86
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
shiboken6==6.8.0.2
pyserial==3.5  # Added for serial communication